End of initial development, read more ShootThis has finally been released. It's been a few months now and we've worked a lot on this game to improve it. We're glad with the outcome. Sure, it could've been done better - but we believe we did our best and honestly, we think it's turned out great, especially for the current state of HTML5 game development. We could've done a better backend and we could've just gone for a regular PC/mobile game, but we just decided we wanted to do this.
What's next? Nobody knows. One thing is for sure - this project was just meant to be a portfolio project. We've done what we wanted to. We might do some improvements here and there but the initial "major" development has been finished at this point.

Whoever ever decides to play this game: we've done our best to document everything for you to be able to run it, at least in 2020/2021.